Upd: New Game 1/3, New Player
New Game additions:
- Added link to create a new player, which appears as a button but it is
actually a link.
- New 25x25 pixels icons for the plus sign and the left, right chevrons
with the last two provided as components for the site's sub-headers.
- Avatar layouts, components and CSS were almost completely re-written
to ensure usability under different circumstances.
- Avatar JS scripts were updated to allow player selection in addition to
avatar selection. This has been a priority -for 3 days now- as both New
Game and New Player pages share almost the same components - yet for
different purposes and unfortunately in different sizes.
- Updated suggestion for Players Model in order to assign a random color
to each New Player profile created. This will require additional support
in the front-end so it will be implemented in a future update.

What remains to be done and has to be discussed:
- Secondary button. Profile should not have the same importance as the
start game button, yet this element is missing from the UI. ~2h
- Delete button. It's too scary to have a delete player in the start
game button. It should be moved to the Settings. ~2h
- Model fixing for Avatars and Players.

What remains to be done and doesn't need to be discussed:
- Trees. Excluding a few responsivity issues their roots have already
been placed and they will grow tomorrow. ~1h

What won't be done:
- Delete button.
- Vue app for the new game screens :/.
- Probably the avatars won't be resized to 88x88 pixels for this page
and they will stay at 100x100 pixels, cause there's no other easy way
to ensure the accessibility of their labels -the actual player names.
Unless of course the trees don't fit.