New: User Header Menu.
The curious case of a long name.

On <x-layout />s is now possible to add a user menu on the header. How?
Simply add :hasUserMenu=true. If you don't need that menu on a layout
then don't even use/set that variable thanks to Blade's love for NULLs.

So, what about that long name? Well...

The dropdown menu indicator -a very nice arrow, created by CSS borders-
is supposed to be attached next to the player's name at the header. If
the name is Κώστας then all is good. But if the name is Πελοποννησιακός
things are getting out of control. If the name is even longer then there
is no Travellers logo anymore. So how did we solve this? So, first we
made sure that the name can't be longer than 100px - that much space is
allocated by design. To do that we had to convert the name's element to
a block. We also had to position it relatively, as the name for some
reason should still be visible when the pulldown is triggered and we
couldn't have z-index otherwise. The same had to be done for the arrow
which is placed ::after the name. Then we moved the arrow somewhere in
between the name and the avatar's circle, as we couldn't possibly link
it to the real width of the text anymore and it was ugly standing by
itself if the name was far shorter e.g. Ιώ. Now, everything fits nicely,
but once again, the names are truncated at about Αλέξανδρ...

Is that really a valid solution? No. Have you considered why on almost
every popular site  these days the avatar is either the person's photo
or their initial letters? On Google for example, you will either see
your photo or your initial letter. Let's say your name is Michael. You
will see an M. That M or your photo is on your avatar at the top right
corner of every Google service. Is there a dropdown indicator there? No.
You know why? If not, read the previous paragraphs. How do people know
to click their initial or their photo? Cause they really like to click
themselves, I guess.

In short, we really need to discuss the name for the player's model.
Until then, the design is ready to handle everything that you will throw
at it.