New: PH-81 components - Select board
As most of the pages in the start-new-game-flow share the same
components, let's start with the only one we don't really need. In
addition, let's add the "disabled" state to boards and introduce on top
of that a brand new design element of ΠΡΟΣΕΧΩΣ. And since we love states
of buttons so much, and all these buttons haven't been used before in
any other section of the site, let's play with their shadows to have
distinct shadows for hover, focus, hover & focus, and of course active.
And since these buttons are wider than the minimum screen-size supported
let's transform them from pills to rectangles on really small screens,
just because we love front-ends.

Also let's make sure that our PNGs have been responsively minimised.
In short, use https://optipng.sourceforge.net/ for lossless minimization
of PNGs, or https://packages.debian.org/bullseye/pngquant for lossy.
Both can be used on Linux, while there are a couple of nice GUI options
for both Windows and macOS.

That's the end of yesterday's notes.