New: PH-86 HTML implementation.
Step 4. Start tutorial or Start game? That's the question. And to do that we had to insert a blank span, make it a block and paint it almost white while keeping the same shady shadows. Uses almost the same components as the ones on the previous steps. Same guidelines for passing selected options. Fully responsive. An option can be disabled via :comingsoon and of course no teaser is added. This is fully responsive as well.