New: PH-94 Settings Tooltips (restructure)
Introducing (responsive) **modal tooltips** for Settings (it shouldn't
really matter if the settings are for a new or an existing player). The
text for these modals is conviniently placed on views/help. There is
still some work that has to be done on the UX, but the structure should
remain the same.

Introducing (bootstrap) tooltips as an alternative for the modals, at
least for the major <legend>s.

Note that neither of these solutions are accessible enough and more ARIA
properties should be introduced before the final version of the game.

This update also resolves some minor issues with sizing, moves the (i)
icon to its proper position, introduces a forward sub-header button and
resolves a few JavaScript issues.