Leftover source images moved to proper path.
Images on source-images are only needed for reference at this point.
They are the original non-optimized versions of images which are already
included in optimised form in the main images folder. They are not moved
to public folder and they should stay in resources for as long as they
are needed (in case of a requested edit).