Fix: PH-168 Header's padding on XXL viewports
On a viewport really close or even wider than the one on which the site was originally created for (1366x768 pixels) the padding at the top of the header (pt-4 or 1.5rem) breaks the illusion of having an app positioned at the exact middle of the screen. Even though 1.5 rem is not much, the top padding should be removed as there is already enough breathing space on top of the header.

This has been achieved in compliance with the Bootstrap breakpoints with the following media-query:

```css
// Remove top padding from header if there's sufficient blank space on top.
@media (min-height: 768px) and (min-width: 1400px) {
    .trvl-header {
        padding-top: 0 !important;
    }
}
```

The !important part is needed to override the existing p-4 that stays true for all other screen sizes (and which is enforced by its own !important thanks to Bootstrap).