Upd: Front-end default keys swap
With this update the front-end is ready for the new default roles for Space and Enter:
Space is now the default key for Navigation, while Enter is now the default key for Selection.

In detail, this update introduces the following changes on the front-end:

Interactive buttons in settingsControls.blade.php & settingsControlsNew.blade.php now have the following default data-attributes:
- Enter as controlAutomaticSelectionButton (overriden by the back-end via the $control_auto_select value, which is parsed by JS)
- Enter as controlManualSelectionButton (overriden by the back-end via the $control_manual_select value, which is parsed by JS)
- Space as controlManualNavigationButton (overriden by the back-end via the control_manual_nav value, which is parsed by JS)

Raw @php defaults in Blade template `switcher.blade.php` with the swapped keys, which are used in case the $switcher array is empty.
JavaScript defaults in `switcher.js` with the swapped keys, which are used in case the $switcher JS array is empty or invalid.

Hopefully, nothing was broken in the process, although some testing will be needed after the relevant back-end update.